Refactored authentication flow
The authentication method now has two scenarios.
- One in which the client provides, besides the username and password,
a tenant name and receives an access object with the service catalog
for that specific tenant
- The other that covers the absence of the tenant name from the client
request by intially requesting a list of tenants for a specific user
and, afterwards, calling the OpenStack API for a full access object
with the first tenant name from the previous list. The finall response
object will include, besides the access object, a list with the previously
requested tenant names.