Added integration for different workers
In order to externalize different operations, like
monitoring, to other services a transport layer was designed
in order to enable communication between the Aurora API and
other services.

The system works by making use of RabbitMQ and the rabbot
node package.

Also a new route was exposed, '/plugins', in order to enable
registration of external plugins that have to consume data
provided by Aurora API.