Added persistance to registered routes
A new functionality was added so that the
gateway reads and writes to routes.yml
information regarding the routing configuration.

The config is loaded on server start, if the file
exists and is valid, and updated each time a new
route is registered on the gateway.

The file structure is similar to what Symfony uses
for routing:
route_name:
	path:
	options: